first refactor attempt
